feat: Implement a splash screen and a feedback form dialog.
